automation/requirements.txt
fpdf2>=2.8.0
Pillow>=11.0.0
